@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <AssemblyName>Moraba.Image.Webp</AssemblyName>
     <RootNamespace>Moraba.Image.Webp</RootNamespace>
     <PackageId>Moraba.Images.Webp.Core</PackageId>
@@ -23,7 +23,7 @@ convert from static path and web path.</Description>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="System.Drawing.Common" Version="4.7.0-preview3.19551.4" />
+    <PackageReference Include="System.Drawing.Common" Version="6.0.0-rc.2.21480.5" />
   </ItemGroup>
 
 </Project>
